refactor: Break up large test files (#237)
* refactor: Break up Tracer unit tests

* refactor: Break up insurance

* chore: Lint TracerMargin

* refactor: Insurance test clean up
