+++ test/unit/Insurance.js
@@ -2,9 +2,7 @@ const { expect } = require("chai")
 const { ethers, getNamedAccounts, deployments } = require("hardhat")
 const { deploy } = deployments
 const { smockit } = require("@eth-optimism/smock")
-const { BigNumber } = require("ethers")
 const zeroAddress = "0x0000000000000000000000000000000000000000"
-const FUNDING_RATE_FACTOR = ethers.utils.parseEther("0.00000570775")
 
 const getCollaterals = async (insurance) => [
     (await insurance.bufferCollateralAmount()).toString(),
@@ -74,22 +72,6 @@ const setup = deployments.createFixture(async () => {
         18
     )
     await testToken.deployed()
-
-    // deploy mock tracer and libs
-    let libBalances = await deploy("Balances", {
-        from: deployer,
-        log: true,
-    })
-
-    let libPerpetuals = await deploy("Perpetuals", {
-        from: deployer,
-        log: true,
-    })
-
-    let libPrices = await deploy("Prices", {
-        from: deployer,
-        log: true,
-    })
 
     // this deploy method is needed for mocking
     const tracerContractFactory = await ethers.getContractFactory(
@@ -183,99 +165,6 @@ describe("Unit tests: Insurance.sol", function () {
                 expect(collateralToken.toString()).to.equal(testToken.address)
             })
             it("emits a pool created event", async () => {})
-        })
-    })
-
-    describe("deposit", async () => {
-        context("when the user does not have enough tokens", async () => {
-            it("reverts", async () => {
-                await expect(
-                    insurance.deposit(ethers.utils.parseEther("1"))
-                ).to.be.revertedWith("ERC20: transfer amount exceeds allowance")
-            })
-        })
-
-        context("when the user has enough tokens", async () => {
-            beforeEach(async () => {
-                await testToken.approve(
-                    insurance.address,
-                    ethers.utils.parseEther("1")
-                )
-                await insurance.deposit(ethers.utils.parseEther("1"))
-            })
-
-            it("mints them pool tokens", async () => {
-                let poolTokenHolding = await insurance.getPoolUserBalance(
-                    accounts[0].address
-                )
-                expect(poolTokenHolding).to.equal(ethers.utils.parseEther("1"))
-            })
-
-            it("increases the collateral holding of the insurance fund", async () => {
-                let collateralHolding = await insurance.publicCollateralAmount()
-                expect(collateralHolding).to.equal(ethers.utils.parseEther("1"))
-            })
-
-            it("pulls in collateral from the tracer market", async () => {
-                let balanceCalls = mockTracer.smocked.getBalance.calls.length
-                expect(balanceCalls).to.equal(1)
-            })
-
-            it("emits an insurance deposit event", async () => {
-                await testToken.approve(
-                    insurance.address,
-                    ethers.utils.parseEther("1")
-                )
-                expect(await insurance.deposit(ethers.utils.parseEther("1")))
-                    .to.emit(insurance, "InsuranceDeposit")
-                    .withArgs(
-                        mockTracer.address,
-                        accounts[0].address,
-                        ethers.utils.parseEther("1")
-                    )
-            })
-        })
-    })
-
-    describe("withdraw", async () => {
-        context("when the user does not have enough pool tokens", async () => {
-            it("reverts", async () => {
-                await expect(
-                    insurance.withdraw(ethers.utils.parseEther("1"))
-                ).to.be.revertedWith("INS: balance < amount")
-            })
-        })
-
-        context("when the user has enough pool tokens", async () => {
-            beforeEach(async () => {
-                // get user tp acquire some pool tokens
-                await testToken.approve(
-                    insurance.address,
-                    ethers.utils.parseEther("2")
-                )
-                await insurance.deposit(ethers.utils.parseEther("2"))
-                // get user to burn some pool tokens
-                await insurance.withdraw(ethers.utils.parseEther("1"))
-            })
-
-            it("burns pool tokens", async () => {
-                let poolTokenHolding = await insurance.getPoolUserBalance(
-                    accounts[0].address
-                )
-                expect(poolTokenHolding).to.equal(ethers.utils.parseEther("1"))
-            })
-
-            it("decreases the collateral holdings of the insurance fund", async () => {
-                let collateralHolding = await insurance.publicCollateralAmount()
-                expect(collateralHolding).to.equal(ethers.utils.parseEther("1"))
-            })
-
-            it("pulls in collateral from the tracer market", async () => {
-                let balanceCalls = mockTracer.smocked.getBalance.calls.length
-                expect(balanceCalls).to.equal(1)
-            })
-
-            it("emits an insurance withdraw event", async () => {})
         })
     })
 
@@ -595,22 +484,6 @@ describe("Unit tests: Insurance.sol", function () {
         })
     })
 
-    describe("getPoolBalance", async () => {
-        context("when called", async () => {
-            it("returns the balance of a user in terms of the pool token", async () => {
-                await testToken.approve(
-                    insurance.address,
-                    ethers.utils.parseEther("2")
-                )
-                await insurance.deposit(ethers.utils.parseEther("2"))
-                let poolBalance = await insurance.getPoolUserBalance(
-                    accounts[0].address
-                )
-                expect(poolBalance).to.equal(ethers.utils.parseEther("2"))
-            })
-        })
-    })
-
     describe("getPoolTarget", async () => {
         context("when called", async () => {
             it("returns 1% of the markets leveraged notional value", async () => {
@@ -622,124 +495,5 @@ describe("Unit tests: Insurance.sol", function () {
                 expect(leveragedNotionalCalls).to.equal(1)
             })
         })
-    })
-
-    describe("getPoolFundingRate", async () => {
-        context("when the leveraged notional value is <= 0", async () => {
-            it("returns 0", async () => {
-                // set leveraged notional value to 0
-                mockTracer.smocked.leveragedNotionalValue.will.return.with(
-                    ethers.utils.parseEther("0")
-                )
-
-                let poolFundingRate = await insurance.getPoolFundingRate()
-                expect(poolFundingRate).to.equal(0)
-            })
-        })
-
-        context("when the pool is greater than the target", async () => {
-            it("returns 0", async () => {
-                // set leveraged notional value to 100
-                mockTracer.smocked.leveragedNotionalValue.will.return.with(
-                    ethers.utils.parseEther("100")
-                )
-
-                let bufferCollateralAmount = "100"
-                let publicCollateralAmount = "100"
-
-                await putCollateral(
-                    mockTracer,
-                    testToken,
-                    insurance,
-                    bufferCollateralAmount,
-                    publicCollateralAmount
-                )
-
-                let poolFundingRate = await insurance.getPoolFundingRate()
-                // poolFundingRate = constant * ratio
-                //                 = ((fund_target - fund_holdings)/fund_holdings) ** 2
-                let ratio = ethers.utils.parseEther("0")
-                let expectedFundingRate = FUNDING_RATE_FACTOR.mul(ratio)
-                    .mul(ratio)
-                    .div(ethers.utils.parseEther("1")) //divide by 1 to simulate WAD math division
-                expect(poolFundingRate).to.equal(expectedFundingRate)
-            })
-        })
-
-        context(
-            "when the leveraged notional value is > 0 and the pool is empty",
-            async () => {
-                it("returns the appropriate one hour funding rate", async () => {
-                    // set leveraged notional value to 400
-                    mockTracer.smocked.leveragedNotionalValue.will.return.with(
-                        ethers.utils.parseEther("400")
-                    )
-
-                    let bufferCollateralAmount = "0"
-                    let publicCollateralAmount = "0"
-
-                    await putCollateral(
-                        mockTracer,
-                        testToken,
-                        insurance,
-                        bufferCollateralAmount,
-                        publicCollateralAmount
-                    )
-
-                    let poolFundingRate = await insurance.getPoolFundingRate()
-                    // poolFundingRate = constant * (ratio ** 2)
-                    //                 = constant * (((fund_target - fund_holdings)/fund_target) ** 2)
-                    //                 = constant * (((4 - 0)/4) ** 2)
-                    //                 = constant * 1
-                    let ratio = ethers.utils.parseEther("1")
-                    let expectedFundingRate = FUNDING_RATE_FACTOR.mul(
-                        ratio
-                    ).div(ethers.utils.parseEther("1")) // should be 0.000570775%
-                    expect(expectedFundingRate).to.equal(
-                        ethers.utils.parseEther("0.00000570775")
-                    )
-
-                    expect(poolFundingRate).to.equal(expectedFundingRate)
-                })
-            }
-        )
-
-        context(
-            "when the leveraged notional value is > 0 and the pool is not empty",
-            async () => {
-                it("returns the appropriate one hour funding rate", async () => {
-                    // set leveraged notional value to 400
-                    mockTracer.smocked.leveragedNotionalValue.will.return.with(
-                        ethers.utils.parseEther("400")
-                    )
-
-                    let bufferCollateralAmount = "1"
-                    let publicCollateralAmount = "1"
-
-                    await putCollateral(
-                        mockTracer,
-                        testToken,
-                        insurance,
-                        bufferCollateralAmount,
-                        publicCollateralAmount
-                    )
-
-                    let poolFundingRate = await insurance.getPoolFundingRate()
-                    // poolFundingRate = constant * (ratio ** 2)
-                    //                 = constant * (((fund_target - fund_holdings)/fund_target) ** 2)
-                    //                 = constant * (((4 - 2)/4) ** 2)
-                    //                 = constant * (0.5 ** 2) = constant * 0.25
-                    let ratio = ethers.utils.parseEther("0.25")
-                    let expectedFundingRate = FUNDING_RATE_FACTOR.mul(
-                        ratio
-                    ).div(ethers.utils.parseEther("1")) // should be 0.000142694063926941%
-                    expect(expectedFundingRate).to.equal(
-                        ethers.utils.parseEther("0.0000014269375") // Approximate due to imprecision
-                    )
-
-                    expect(poolFundingRate).to.equal(expectedFundingRate)
-                })
-            }
-        )
     })
 })
